moved dialog to seperate file
